requirements.txt
Faker
mysql-connector-python
neo4j